add Culture to Step Attributes
split Attributes.cs to multiple files
