+++ TechTalk.SpecFlow/StepArgumentTransformationAttribute.cs
@@ -1,0 +1,20 @@
+using System;
+
+namespace TechTalk.SpecFlow
+{
+    [AttributeUsage(AttributeTargets.Method, AllowMultiple = true)]
+    public class StepArgumentTransformationAttribute : Attribute
+    {
+        public string Regex { get; set; }
+
+        public StepArgumentTransformationAttribute(string regex)
+        {
+            Regex = regex;
+        }   
+        
+        public StepArgumentTransformationAttribute()
+        {
+            Regex = null;
+        }
+    }
+}
