+++ TechTalk.SpecFlow/TechTalk.SpecFlow.csproj
@@ -157,6 +157,7 @@
     <Compile Include="Bindings\StepDefinitionRegexCalculator.cs" />
     <Compile Include="BoDi\BoDi.cs" />
     <Compile Include="EnvironmentVariableNames.cs" />
+    <Compile Include="HookAttributes.cs" />
     <Compile Include="Infrastructure\BindingInstanceResolver.cs" />
     <Compile Include="Infrastructure\ContextManager.cs" />
     <Compile Include="Infrastructure\ContextManagerExtensions.cs" />
@@ -192,6 +193,8 @@
     <Compile Include="Plugins\RuntimePluginParameters.cs" />
     <Compile Include="ProgrammingLanguage.cs" />
     <Compile Include="ScenarioStepContext.cs" />
+    <Compile Include="ScopeAttributes.cs" />
+    <Compile Include="StepArgumentTransformationAttribute.cs" />
     <Compile Include="StepInfo.cs" />
     <Compile Include="Steps.cs" />
     <Compile Include="Bindings\StepArgumentTransformationBinding.cs" />
